list_products/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
pytest==7.4.3
pytest-assume==2.4.3
flask_cors
mysql-connector-python==8.0.31
jaeger-client==4.6.0
opentracing==2.4.0
six==1.16.0
threadloop==1.0.2
thrift==0.13.0
tornado==6.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger